+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Function `get_key_fn` to customize sorting using patterns.
 - Option `on_error` to `load_yaml` to ignore yaml scanner and parser exceptions.
+- Function `ranks` to get the ranking values in a tensor.
+- Function `sort_with_patterns` to sort values using string-based patterns.
 
 ### Changed
 - Rename `is_pickle_root` to `is_packed_root` (old name was kept fpr backward compatibility).